+++ datacube_zarr/utils/chunk.py
@@ -73,10 +73,8 @@ def validate_chunks(
     if chunks is None:
         chunks = {}
 
-    # Check all dimensions are valid
-    invalid_dims = [d for d in chunks if d not in dims]
-    if invalid_dims:
-        raise ValueError(f"Invalid chunking dim(s) specified: {invalid_dims}.")
+    # Remove chunk specifications for dimensions not present
+    chunks = {k: v for k, v in chunks.items() if k in dims}
 
     # Check chunk values are valid
     def _is_valid_chunks(c: Union[int, str]) -> bool:
@@ -123,7 +121,7 @@ def chunk_dataset(
     else:
         chunk_size_bytes = target_mb * (1024 ** 2) * compression_ratio
         for name, da in ds.data_vars.items():
-            da_chunk_total = chunk_size_bytes / da.dtype.itemsize
+            da_chunk_total = int(chunk_size_bytes // da.dtype.itemsize)
             da_chunks = calculate_auto_chunk_sizes(da.sizes, chunks, da_chunk_total)
             logger.debug(
                 f"Auto chunking array {name} with: {da_chunks} (target_mb={target_mb}, "