+++ integration_tests/test_zarrify_prepare_index.py
@@ -7,7 +7,7 @@ from click.testing import CliRunner
 from datacube.api.core import Datacube
 from datacube.index.hl import Doc2Dataset
 
-from datacube_zarr.tools.zarrify import main as zarrify
+from datacube_zarr.tools.zarrify import cli as zarrify
 from examples.eo3.eo3prepare_usgs_espa_ls8c1_l2_zarr import main as prepare_zarr_ls8
 from examples.prepare_zarr_ls5 import main as prepare_zarr_ls5
 from integration_tests.conftest import TEST_DATA_LS8
@@ -181,9 +181,6 @@ def test_zarrify_prepare_index_ls8_eo3(
             output_crs=output_crs,
             resolution=resolution,
         )
-
-        print(data_tiff)
-        print(data_zarr)
 
         # compare datasets
         assert data_zarr.equals(data_tiff)