+++ tests/test_driver.py
@@ -1,22 +1,20 @@
 '''Unit tests for the datacube_zarr.driver module.'''
+from pathlib import Path
 from random import random, sample
 
 import pytest
 import numpy as np
-from datacube.drivers import reader_drivers, writer_drivers
+from datacube.drivers import reader_drivers
 from datacube.storage import BandInfo
+from datacube.testutils.iodriver import open_reader
 from mock import MagicMock
 
 from datacube_zarr.driver import (
     ZarrDataSource,
     ZarrReaderDriver,
-    ZarrWriterDriver,
     reader_driver_init,
     uri_split,
-    writer_driver_init,
 )
-
-from .utils import _check_zarr_files, _load_dataset
 
 SPECTRAL_DEFINITION = {
     'wavelength': sorted(sample(range(380, 750), 150)),
@@ -31,13 +29,6 @@ def test_reader_drivers():
     available_drivers = reader_drivers()
     assert isinstance(available_drivers, list)
     assert 'zarr' in available_drivers
-
-
-def test_writer_drivers():
-    '''Check the zarr writer driver is found by Datacube.'''
-    available_drivers = writer_drivers()
-    for name in ('zarr file', 'zarr s3'):
-        assert name in available_drivers
 
 
 def test_zarr_netcdf_driver_import():
@@ -57,11 +48,12 @@ def test_datasource(request, dataset, dataset_fixture):
 
     Data is saved to file and opened by the data source."""
     odc_dataset_ = request.getfixturevalue(dataset_fixture)
+
     group_name = list(dataset.keys())[0]
     source = ZarrDataSource(BandInfo(odc_dataset_, group_name))
     with source.open() as band_source:
         ds = band_source.read()
-        assert np.array_equal(ds, dataset[group_name].values[0, ...])
+        assert np.array_equal(ds.squeeze(), dataset[group_name].values.squeeze())
 
 
 def test_datasource_empty_band_info(dataset):
@@ -83,12 +75,12 @@ def test_datasource_wrong_protocol(dataset):
     assert str(excinfo.value) == 'Expected file:// or s3:// url'
 
 
-def test_datasource_no_timeslice(dataset):
+def test_datasource_read_window(dataset):
     '''Test the ZarrDataSource.BandDataSource.'''
     group_name = list(dataset.keys())[0]
 
     band_source = ZarrDataSource.BandDataSource(
-        dataset, group_name, None, dataset[group_name].nodata
+        dataset, group_name, dataset[group_name].nodata
     )
     assert band_source.crs == dataset.crs
     assert band_source.transform == dataset[group_name].affine
@@ -96,20 +88,36 @@ def test_datasource_no_timeslice(dataset):
     assert band_source.shape == dataset[group_name].shape[1:]
 
     ds2 = band_source.read()
-    assert np.array_equal(ds2, dataset[group_name].values[0, ...])
+    assert np.array_equal(ds2, dataset[group_name].values)
 
-    ds3 = band_source.read(((30, 50), (30, 50)))
+    ds3 = band_source.read((0, (30, 50), (30, 50)))
     assert np.array_equal(ds3, dataset[group_name][0, 30:50, 30:50].values)
 
 
-def test_datasource_bad_time_index(dataset):
+def test_datasource_read_outshape(dataset, tmpdir):
+    """Test that reads with `out_shape` specified match the rio driver."""
+    group_name = list(dataset.keys())[0]
+    z_src = ZarrDataSource.BandDataSource(dataset, group_name, dataset[group_name].nodata)
+
+    # Create a ODC `RIOReader` reader from the original test tiff file
+    tif = next(Path(tmpdir).iterdir())
+    t_src = open_reader(tif)
+    for out_shape in (None, (0, 1), (2, 2), (4, 7), (16, 24), (32, 30), (143, 52)):
+        t_data = t_src.read(out_shape=out_shape).result()
+        z_data = z_src.read((0,), out_shape=out_shape)
+        np.testing.assert_array_equal(z_data, t_data)
+
+
+def test_datasource_bad_window(dataset):
     '''Test the ZarrDataSource.BandDataSource with an invalid time index.'''
     group_name = list(dataset.keys())[0]
+    band_source = ZarrDataSource.BandDataSource(
+        dataset, group_name, dataset[group_name].nodata
+    )
+
     with pytest.raises(IndexError) as excinfo:
-        ZarrDataSource.BandDataSource(
-            dataset, group_name, dataset.time.size + 1, dataset[group_name].nodata
-        )
-    assert str(excinfo.value) == 'band_idx 1 out of range (nbands=1)'
+        band_source.read((1,))
+    assert str(excinfo.value) == 'index 1 is out of bounds for axis 0 with size 1'
 
 
 def test_datasource_no_time_slice(dataset):
@@ -117,9 +125,7 @@ def test_datasource_no_time_slice(dataset):
     group_name = list(dataset.keys())[0]
     dataset = dataset.drop_sel(time=dataset.time.values)
     with pytest.raises(ValueError) as excinfo:
-        ZarrDataSource.BandDataSource(
-            dataset, group_name, None, dataset[group_name].nodata
-        )
+        ZarrDataSource.BandDataSource(dataset, group_name, dataset[group_name].nodata)
     assert str(excinfo.value) == 'Dataset has 0 bands.'
 
 
@@ -128,7 +134,7 @@ def test_datasource_no_nodata(dataset):
     group_name = list(dataset.keys())[0]
     dataset[group_name].attrs.pop('nodata', None)
     with pytest.raises(ValueError) as excinfo:
-        ZarrDataSource.BandDataSource(dataset, group_name, dataset.time.size, None)
+        ZarrDataSource.BandDataSource(dataset, group_name, None)
     assert str(excinfo.value) == 'nodata not found in dataset and product definition'
 
 
@@ -146,8 +152,8 @@ def test_datasource_stacked_nodata(dataset):
     '''Test the ZarrDataSource.BandDataSource with nodata as a list.'''
     group_name = list(dataset.keys())[0]
     dataset.aa.attrs['nodata'] = [-9999]
-    band_source = ZarrDataSource.BandDataSource(dataset, group_name, None, None)
-    assert band_source._nodata == -9999
+    band_source = ZarrDataSource.BandDataSource(dataset, group_name, None)
+    assert band_source.nodata == -9999
 
 
 @pytest.mark.parametrize("uri,split_uri", uri_split_test_params)
@@ -181,89 +187,4 @@ def test_zarr_reader_driver(dataset, odc_dataset):
     source = reader.new_datasource(BandInfo(odc_dataset, group_name))
     with source.open() as band_source:
         ds = band_source.read()
-        assert np.array_equal(ds, dataset[group_name].values[0, ...])
-
-
-def test_zarr_writer_driver():
-    '''Check aliases, format and uri_scheme for the writer.'''
-    writer = writer_driver_init()
-    assert isinstance(writer, ZarrWriterDriver)
-    assert writer.aliases == ['zarr file', 'zarr s3']
-    assert writer.format == 'zarr'
-
-
-def test_writer_driver_mk_uri():
-    '''Check mk_uri for the writer for supported aliases'''
-    writer_driver = ZarrWriterDriver()
-
-    # Test 'zarr file' driver alias
-    file_path = '/path/to/my_file.zarr'
-    driver_alias = 'zarr file'
-    storage_config = {'driver': driver_alias}
-    file_uri = writer_driver.mk_uri(file_path=file_path, storage_config=storage_config)
-    assert file_uri == f'file://{file_path}'
-
-    # Test 'zarr s3' driver alias
-    file_path = 'bucket/path/to/my_file.zarr'
-    driver_alias = 'zarr s3'
-    storage_config = {'driver': driver_alias}
-    file_uri = writer_driver.mk_uri(file_path=file_path, storage_config=storage_config)
-    assert file_uri == f's3://{file_path}'
-
-    # Test unknown driver alias
-    file_path = 'bucket/path/to/my_file.zarr'
-    driver_alias = 'unknown alias'
-    storage_config = {'driver': driver_alias}
-    with pytest.raises(ValueError) as excinfo:
-        file_uri = writer_driver.mk_uri(
-            file_path=file_path, storage_config=storage_config
-        )
-    assert str(excinfo.value) == f'Unknown driver alias: {driver_alias}'
-
-
-def test_zarr_file_writer_driver_save(uri, fixed_chunks, data, s3):
-    '''Test the `write_dataset_to_storage` method.'''
-    name = 'array1'
-    writer = ZarrWriterDriver()
-    ds_in = data.to_dataset(name=name)
-    writer.write_dataset_to_storage(
-        dataset=ds_in.copy(),
-        file_uri=uri,
-        storage_config={'chunking': fixed_chunks['input']},
-    )
-    _check_zarr_files(data, uri, name, fixed_chunks, s3)
-
-    # Load and check data
-    ds_out = _load_dataset(uri)
-    assert ds_in.equals(ds_out)  # Compare values only
-
-
-def test_zarr_file_writer_driver_data_corrections(uri, fixed_chunks, data):
-    '''Test dataset key corrections applied by `write_dataset_to_storage`.'''
-    name = 'array1'
-    writer = ZarrWriterDriver()
-    ds_in = data.to_dataset(name=name)
-    # Assign target keys: spectral definition and coords attributes
-    ds_in.array1.attrs['spectral_definition'] = SPECTRAL_DEFINITION
-    coords = {dim: [1] * size for dim, size in ds_in.dims.items()}
-    ds_in = ds_in.assign_coords(coords)
-    for coord_name in ds_in.coords:
-        ds_in.coords[coord_name].attrs['units'] = 'Fake unit'
-    writer.write_dataset_to_storage(
-        dataset=ds_in.copy(),  # The copy should be corrected
-        file_uri=uri,
-        storage_config={'chunking': fixed_chunks['input']},
-    )
-    # Load and check data has been corrected
-    ds_out = _load_dataset(uri)
-    assert ds_in.equals(ds_out)  # Values only
-    for key, value in SPECTRAL_DEFINITION.items():
-        # spectral defs attributes should now start with 'dc_'
-        assert ds_out.array1.attrs[f'dc_spectral_definition_{key}'] == value  # attrs
-    for coord_name in ds_in.coords:
-        assert ds_out.coords[coord_name].equals(ds_in.coords[coord_name])
-        for attr, val in ds_in.coords[coord_name].attrs.items():
-            # units attribute should now start with 'dc_'
-            if attr == 'units':
-                attr = 'dc_units'
-            assert ds_out.coords[coord_name].attrs[attr] == val
+        assert np.array_equal(ds, dataset[group_name].values)